fix(learn): reset update state across lesson contexts
Clear pending off-screen update indicators when chat or lesson context changes without disturbing same-lesson retries.
Split catalog and lesson-content lifecycles out of LearnView so each state owner remains bounded and independently testable.
